+++ src/organisms/ContestTabPage/index.js
@@ -1,0 +1,3 @@
+import ContestTabPageContainer from "./ContestTabPageContainer";
+
+export default ContestTabPageContainer;
